+++ flutter/lib/desktop/pages/remote_page.dart
@@ -139,6 +139,7 @@ class _RemotePageState extends State<RemotePage>
     _ffi.ffiModel.updateEventListener(sessionId, widget.id);
     if (!isWeb) bind.pluginSyncUi(syncTo: kAppTypeDesktopRemote);
     _ffi.qualityMonitorModel.checkShowQualityMonitor(sessionId);
+    _ffi.daxianStatusModel.checkShowDaxianStatusMonitor(sessionId);
     _ffi.dialogManager.loadMobileActionsOverlayVisible();
     WidgetsBinding.instance.addPostFrameCallback((_) {
       // Session option should be set after models.dart/FFI.start
@@ -541,7 +542,11 @@ class _RemotePageState extends State<RemotePage>
         top: 10,
         right: 10,
         child: _buildRawTouchAndPointerRegion(
-            QualityMonitor(_ffi.qualityMonitorModel), null, null),
+          RemoteStatusMonitors(
+              _ffi.qualityMonitorModel, _ffi.daxianStatusModel),
+          null,
+          null,
+        ),
       ),
     );
     return Stack(