+++ libs/hbb_common/src/config.rs
@@ -294,6 +294,8 @@ pub struct PeerConfig {
     #[serde(flatten)]
     pub show_quality_monitor: ShowQualityMonitor,
     #[serde(flatten)]
+    pub show_daxian_status_monitor: ShowDaxianStatusMonitor,
+    #[serde(flatten)]
     pub follow_remote_cursor: FollowRemoteCursor,
     #[serde(flatten)]
     pub follow_remote_window: FollowRemoteWindow,
@@ -378,6 +380,7 @@ impl Default for PeerConfig {
             disable_clipboard: Default::default(),
             enable_file_copy_paste: Default::default(),
             show_quality_monitor: Default::default(),
+            show_daxian_status_monitor: Default::default(),
             follow_remote_cursor: Default::default(),
             follow_remote_window: Default::default(),
             keyboard_mode: Default::default(),
@@ -1642,6 +1645,12 @@ serde_field_bool!(
     "ShowQualityMonitor::default_show_quality_monitor"
 );
 serde_field_bool!(
+    ShowDaxianStatusMonitor,
+    "show_daxian_status_monitor",
+    default_show_daxian_status_monitor,
+    "ShowDaxianStatusMonitor::default_show_daxian_status_monitor"
+);
+serde_field_bool!(
     DisableAudio,
     "disable_audio",
     default_disable_audio,
@@ -1947,6 +1956,7 @@ impl UserDefaultConfig {
             keys::OPTION_CUSTOM_IMAGE_QUALITY => self.get_num_string(key, 50.0, 10.0, 0xFFF as f64),
             keys::OPTION_CUSTOM_FPS => self.get_num_string(key, 30.0, 5.0, 120.0),
             keys::OPTION_ENABLE_FILE_COPY_PASTE => self.get_string(key, "Y", vec!["", "N"]),
+            keys::OPTION_SHOW_DAXIAN_STATUS_MONITOR => self.get_string(key, "Y", vec!["", "N"]),
             keys::OPTION_TRACKPAD_SPEED => self.get_num_string(key, 100, 10, 1000),
             _ => self
                 .get_after(key)
@@ -2413,6 +2423,7 @@ pub mod keys {
     pub const OPTION_FOLLOW_REMOTE_WINDOW: &str = "follow_remote_window";
     pub const OPTION_ZOOM_CURSOR: &str = "zoom-cursor";
     pub const OPTION_SHOW_QUALITY_MONITOR: &str = "show_quality_monitor";
+    pub const OPTION_SHOW_DAXIAN_STATUS_MONITOR: &str = "show_daxian_status_monitor";
     pub const OPTION_DISABLE_AUDIO: &str = "disable_audio";
     pub const OPTION_ENABLE_REMOTE_PRINTER: &str = "enable-remote-printer";
     pub const OPTION_ENABLE_FILE_COPY_PASTE: &str = "enable-file-copy-paste";
@@ -2566,6 +2577,7 @@ pub mod keys {
         OPTION_FOLLOW_REMOTE_WINDOW,
         OPTION_ZOOM_CURSOR,
         OPTION_SHOW_QUALITY_MONITOR,
+        OPTION_SHOW_DAXIAN_STATUS_MONITOR,
         OPTION_DISABLE_AUDIO,
         OPTION_ENABLE_FILE_COPY_PASTE,
         OPTION_DISABLE_CLIPBOARD,